Do not allow variables to leak into global scope
Add 'setlocal' to the top of the Batch script to ensure the script's
internal variables do not leak out into the global scope.

This was causing a bug with the Restore GUI functionality when used
from the command-prompt. Between multiple uses of the 'rawaccel-switch'
command, the GUI was sometimes being launched when it shouldn't have
been, simply because it had been launched on a prior occasion. This was
due to a globally-scoped variable remaining defined until the end of
the entire command-prompt session. Using 'setlocal' prevents this.